2h06min ade counter in navbar and Lifting state up
