+++ my-app/src/App.js
@@ -2,17 +2,55 @@ import NavBar from './components/navbar';
 import './App.css';
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import Counters from './components/counters';
-import React from 'react';
+import React, { Component } from 'react';
 
-function App() {
+class App extends Component {
+  state = { 
+    counters: [
+        {id:1, value:0},
+        {id:2, value:0},
+        {id:3, value:0},
+        {id:4, value:0},
+    ]
+ };
+
+handleIncrement = counter => {
+const counters = [...this.state.counters];
+const index = counters.indexOf(counter);
+counters[index] = {...counter};
+counters[index].value++;
+this.setState({counters});
+}
+
+handleReset = () => {
+const counters = this.state.counters.map( c => {
+c.value=0;
+return c;
+})
+this.setState({})
+}
+
+ handleDelete = (counterId) => {
+    const counters = this.state.counters.filter(c => c.id !== counterId);
+    this.setState({counters});
+ }
+  render () {
   return (
-    <React.Fragment>
-    <NavBar/>
-    <main className='container'>
-      <Counters />
-    </main>
-    </React.Fragment>
-  )
+      <React.Fragment>
+      <NavBar 
+        totalCounters={this.state.counters.filter(c => c.value > 0).length}
+      />
+      <main className='container'>
+        <Counters 
+        counters={this.state.counters}
+        onReset={this.handleReset}
+        onDelete={this.handleDelete}
+        onIncremet={this.handleIncrement}
+        />
+      </main>
+      </React.Fragment>
+    )
+  }
 }
 
 export default App;
